dynamic_pricing/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.1.0
lightgbm==4.4.0
catboost==1.2.5
scipy==1.13.1
openpyxl==3.1.4
pyarrow==16.1.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.3
redis==5.0.6
kafka-python==2.0.2
streamlit==1.36.0
plotly==5.22.0
streamlit-autorefresh==1.0.1
python-dotenv==1.0.1
joblib==1.4.2
pytest==8.2.2
httpx==0.27.0
tqdm==4.66.4
matplotlib==3.8.4
seaborn==0.13.2
